+++ peachStore/src/main/java/peachstore/admin/controller/PaymentController.java
@@ -35,159 +35,4 @@ import peachstore.service.orderreceipt.OrderReceiptService;
 @RequiredArgsConstructor
 @Controller
 public class PaymentController {
-
-	private final TossPaymentService tossPaymentService;
-	private final OrderReceiptService orderReceiptService;
-
-	private final TosspaymentDAO tosspaymentDAO;
-
-	/*
-	 * 결제 입장
-	 */
-	@GetMapping("/payment/start")
-	public String paymentPage(Model model, int orderReceiptId, Long amount, String userName) {
-		String successUrl = "http://localhost:8888/admin/payment/success-handler?orderReceiptId=" + orderReceiptId;
-		String failUrl = "http://localhost:8888/admin/payment/fail";
-		String orderId = String.valueOf(System.currentTimeMillis());
-
-		log.debug("paymentPage orderId - {}, orderReceiptId - {}, amount - {}, userName - {}", orderId, orderReceiptId,
-				amount, userName);
-		
-		model.addAttribute("orderReceiptId", orderReceiptId);
-		model.addAttribute("orderId", orderId);
-		model.addAttribute("amount", amount);
-		model.addAttribute("customerName", userName);
-		model.addAttribute("successUrl", successUrl);
-		model.addAttribute("failUrl", failUrl);
-		model.addAttribute("orderName", "주문번호 " + orderId + " 결제");
-
-		return "/payment/payment";
-	}
-
-	/**
-	 * 결제 완료 중간 단계로 이동
-	 */
-	@GetMapping("/payment/success-handler")
-	public String successHandlerPage(@RequestParam String paymentKey, @RequestParam String orderId,
-			@RequestParam long amount, @RequestParam int orderReceiptId, Model model) {
-		model.addAttribute("paymentKey", paymentKey);
-		model.addAttribute("orderId", orderId);
-		model.addAttribute("amount", amount);
-		model.addAttribute("orderReceiptId", orderReceiptId);
-
-		return "/payment/success-handler";
-	}
-
-	/**
-	 * 결제 정보 db 저장
-	 * 
-	 * @param req
-	 * @return
-	 */
-	@PostMapping("/payment/confirm")
-	public ResponseEntity<TossConfirmResponse> successHandlerPageConfirm(@RequestBody ConfirmPaymentRequest req) {
-
-		log.debug("successHandlerPageConfirm 호출");
-		// 토스 결제 승인 요청
-		TossConfirmResponse response = tossPaymentService.confirmPayment(req.getPaymentKey(), req.getOrderId(),
-				req.getAmount());
-
-		// 받은 정보로 Tosspayment 객체 세팅
-		Tosspayment tosspayment = new Tosspayment();
-		tosspayment.setTossPaymentKey(req.getPaymentKey());
-		tosspayment.setTossOrderId(req.getOrderId());
-		tosspayment.setTotalAmount(req.getAmount());
-		tosspayment.setTossPaymentMethod(response.getMethod());
-		tosspayment.setTossPaymentStatus(response.getStatus());
-		tosspayment.setApprovedAt(response.getApprovedAt().toLocalDateTime()); // OffsetDateTime -> LocalDateTime 변환
-		tosspayment.setRequestedAt(response.getRequestedAt().toLocalDateTime());
-		OrderReceipt orderReceipt = orderReceiptService.selectById(req.getOrderReceiptId());
-		tosspayment.setOrderReceipt(orderReceipt);
-
-		// DB 저장
-		tosspaymentDAO.insert(tosspayment);
-
-		// 성공 결과 반환
-		return ResponseEntity.ok(response);
-	}
-
-	@GetMapping("/payment/success")
-	public String successPage() {
-		return "/payment/success";
-	}
-
-	/**
-	 * 결제 실패 페이지
-	 */
-	@GetMapping("/payment/fail")
-	public String failPage(@RequestParam Map<String, String> params, Model model) {
-		model.addAllAttributes(params);
-		return "/payment/fail";
-	}
-
-	/**
-	 * 토스에 결제 승인받기
-	 * 
-	 * @param confirmPaymentRequest
-	 * @return
-	 * @throws Exception
-	 */
-	private String getAuthorizations() {
-		log.debug("getAuthorizations");
-		String secretKey = "test_sk_zXLkKEypNArWmo50nX3lmeaxYG5R"; // 토스 Secret Key
-		String auth = secretKey + ":";
-		return "Basic " + Base64.getEncoder().encodeToString(auth.getBytes());
-	}
-
-	/**
-	 * 토스에게 결제 승인 요청
-	 */
-	public String requestConfirm(ConfirmPaymentRequest confirmPaymentRequest) throws IOException {
-		log.debug("requestConfirm");
-		String tossOrderId = confirmPaymentRequest.getOrderId();
-		long amount = confirmPaymentRequest.getAmount();
-		String tossPaymentKey = confirmPaymentRequest.getPaymentKey();
-
-		ObjectMapper objectMapper = new ObjectMapper();
-
-		JsonNode requestObj = objectMapper.createObjectNode().put("orderId", tossOrderId).put("amount", amount)
-				.put("paymentKey", tossPaymentKey);
-
-		String requestBody = objectMapper.writeValueAsString(requestObj);
-
-		HttpPost post = new HttpPost("https://api.tosspayments.com/v1/payments/confirm");
-		post.setHeader("Authorization", getAuthorizations()); // getAuthorizations() 메서드가 Authorization 헤더 문자열 반환
-		post.setHeader("Content-Type", "application/json");
-		post.setEntity(new StringEntity(requestBody, "UTF-8"));
-
-		try (CloseableHttpClient client = HttpClients.createDefault();
-				CloseableHttpResponse response = client.execute(post)) {
-
-			String responseBody = EntityUtils.toString(response.getEntity(), "UTF-8");
-			return responseBody;
-		}
-	}
-
-	/**
-	 * 결제 취소 요청
-	 */
-	public String cancelPayment(String paymentKey, String cancelReason) throws IOException {
-		log.debug("cancelPayment");
-		String cancelUrl = "https://api.tosspayments.com/v1/payments/" + paymentKey + "/cancel";
-
-		// JSON 본문 구성
-		String requestBody = "{\"cancelReason\": \"" + cancelReason + "\"}";
-
-		HttpPost post = new HttpPost(cancelUrl);
-		post.setHeader("Authorization", getAuthorizations());
-		post.setHeader("Content-Type", "application/json");
-		post.setEntity(new StringEntity(requestBody, "UTF-8"));
-
-		try (CloseableHttpClient client = HttpClients.createDefault();
-				CloseableHttpResponse response = client.execute(post)) {
-
-			String responseBody = EntityUtils.toString(response.getEntity(), "UTF-8");
-			return responseBody;
-		}
-	}
 }
