+++ peachStore/src/main/java/peachstore/repository/orderreceipt/OrderReceiptDAO.java
@@ -15,4 +15,11 @@ public interface OrderReceiptDAO {
 	public OrderReceipt selectById(int orderReceiptId);
 	
 	public List  completedList(OrderReceipt orderReceipt);
+	
+	/**
+	 * 주문내역 생성
+	 * @param orderReceipt
+	 * @return
+	 */
+	public int insert(OrderReceipt orderReceipt);
 }
