+++ peachStore/src/main/java/peachstore/repository/orderreceipt/OrderReceiptDAOImpl.java
@@ -50,4 +50,11 @@ public class OrderReceiptDAOImpl implements OrderReceiptDAO{
 		return sqlSessionTemplate.selectOne("OrderReceipt.select", orderReceiptId);
 	}
 
+	@Override
+	public int insert(OrderReceipt orderReceipt) {
+		int result = sqlSessionTemplate.insert("OrderReceipt.insert", orderReceipt);
+		log.debug("insert - count: {}", result);
+		return result;
+	}
+
 }
