+++ peachStore/src/main/webapp/WEB-INF/views/shop/payment/success.jsp
@@ -1,0 +1,12 @@
+
+ 
+<%@ page contentType="text/html; charset=UTF-8" %><%@ page contentType="text/html; charset=UTF-8" %>
+<html>
+<head>
+<title>결제 성공</title>
+</head>
+<body>
+<h1>결제가 성공적으로 완료되었습니다!</h1>
+<p>주문 번호: ${orderReceiptId}</p>
+</body>
+</html>
